GET friends, removed test values
